test: split code and add tests
